+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/InventoryService.java
@@ -394,7 +394,6 @@ public class InventoryService {
                 product.getDescription(),
                 diff,
                 avgPrice,
-                avgPrice,
                 product.getUnit(),
                 stockMove,
                 StockMoveLineService.TYPE_NULL,