+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/exception/IExceptionMessage.java
@@ -79,13 +79,6 @@ public interface IExceptionMessage {
   static final String INCOMING_STOCK_MOVE_INVOICE_EXISTS = /*$$(*/
       "An invoice not canceled already exists for the incoming stock move %s" /*)*/;
 
-  /** Stock move line service */
-  static final String STOCK_MOVE_MISSING_SALE_ORDER = /*$$(*/
-      "Missing link to sale order line (from sale order id = %s) for stock move line %s" /*)*/;
-
-  static final String STOCK_MOVE_MISSING_PURCHASE_ORDER = /*$$(*/
-      "Missing purchase order with id %s for stock move line %s" /*)*/;
-
   /** Batch Invoicing */
   static final String BATCH_INVOICING_1 = /*$$(*/ "Subscription invoice generation report :" /*)*/;